Refactor rendering API and update framebuffer formats
Modifications to `OpenGLRendererAPI.cpp` and `VulkanFramebuffer.cpp` focus on the Kerberos namespace. The `glViewport` function now casts parameters to `int` for improved type safety, and several OpenGL features are commented out, indicating a potential shift in rendering strategy. A new case for `FramebufferTextureFormat::RED_INTEGER` is added in `VulkanFramebuffer.cpp`, expanding framebuffer format options. Additionally, multiple shader files have been updated, reflecting changes in shader code or optimizations, as seen in the binary data changes in the `.spv` files. These updates enhance rendering capabilities and shader management within the Kerberos graphics framework.

+++ Kerberos/src/Platform/OpenGL/OpenGLRendererAPI.cpp
@@ -45,12 +45,16 @@ namespace Kerberos
 		glEnable(GL_POLYGON_OFFSET_FILL);
 		glPolygonOffset(1.0f, 1.0f);
 		//glEnable(GL_LINE_SMOOTH);
+
+		//glEnable(GL_CULL_FACE);
+
+		/*glEnable(GL_FRAMEBUFFER_SRGB);*/
 	}
 
 	void OpenGLRendererAPI::SetViewport(const uint32_t x, const uint32_t y, const uint32_t width,
 		const uint32_t height) 
 	{
-		glViewport(x, y, width, height);
+		glViewport(static_cast<int>(x), static_cast<int>(y), static_cast<int>(width), static_cast<int>(height));
 	}
 
 	void OpenGLRendererAPI::SetClearColor(const glm::vec4& color)
